Porter stemmer algorithm for OCaml
Import of some old code of mine, originally written in 2003.  I'm not
very sure if this works with newer versions of OCaml, or if it's
functional in any way at all.

Licensing is also a problem.  The original stemmer.c file is under
copyright, so once this actually works, the C implementation should
be removed.